HCPPS-1426 Modify azure peering connection test file to include an internal test with configurable hub/spoke parameters
