+++ internal/providersdkv2/resource_azure_peering_connection_test.go
@@ -16,93 +16,115 @@ import (
 )
 
 var (
-	uniqueAzurePeeringTestID  = fmt.Sprintf("hcp-provider-test-%s", time.Now().Format("200601021504"))
-	subscriptionID            = os.Getenv("ARM_SUBSCRIPTION_ID")
-	tenantID                  = os.Getenv("ARM_TENANT_ID")
-	testAccAzurePeeringConfig = fmt.Sprintf(`
-	  provider "azurerm" {
-		features {}
-	  }
-
-	  resource "hcp_hvn" "test" {
-		hvn_id         = "%[1]s"
-		cloud_provider = "azure"
-		region         = "eastus"
-		cidr_block     = "172.25.16.0/20"
-	  }
-
-	  // This resource initially returns in a Pending state, because its application_id is required to complete acceptance of the connection.
-	  resource "hcp_azure_peering_connection" "peering" {
-		hvn_link                 = hcp_hvn.test.self_link
-		peering_id               = "%[1]s"
-		peer_vnet_name           = azurerm_virtual_network.vnet.name
-		peer_subscription_id     = "%[2]s"
-		peer_tenant_id           = "%[3]s"
-		peer_resource_group_name = azurerm_resource_group.rg.name
-		peer_vnet_region         = "eastus"
-
-		allow_forwarded_traffic = false
-		use_remote_gateways     = false
-	  }
-
-	  // This data source is the same as the resource above, but waits for the connection to be Active before returning.
-	  data "hcp_azure_peering_connection" "peering" {
-		hvn_link                 = hcp_hvn.test.self_link
-		peering_id               = hcp_azure_peering_connection.peering.peering_id
-		wait_for_active_state    = true
-	  }
-
-	  // The route depends on the data source, rather than the resource, to ensure the peering is in an Active state.
-	  resource "hcp_hvn_route" "route" {
-		hvn_route_id = "%[1]s"
-		hvn_link = hcp_hvn.test.self_link
-		destination_cidr = "172.31.0.0/16"
-		target_link = data.hcp_azure_peering_connection.peering.self_link
-	  }
-
-	  resource "azurerm_resource_group" "rg" {
-		name     = "%[1]s"
-		location = "East US"
-	  }
-
-	  resource "azurerm_virtual_network" "vnet" {
-		name                = "%[1]s"
-		location            = azurerm_resource_group.rg.location
-		resource_group_name = azurerm_resource_group.rg.name
-
-		address_space = [
-		  "10.0.0.0/16"
-		]
-	  }
-
-	  resource "azuread_service_principal" "principal" {
-		application_id = hcp_azure_peering_connection.peering.application_id
-	  }
-
-	  resource "azurerm_role_definition" "definition" {
-		name  = "%[1]s"
-		scope = azurerm_virtual_network.vnet.id
-
-		assignable_scopes = [
-		  azurerm_virtual_network.vnet.id
-		]
-
-		permissions {
-		  actions = [
-			"Microsoft.Network/virtualNetworks/peer/action",
-			"Microsoft.Network/virtualNetworks/virtualNetworkPeerings/read",
-			"Microsoft.Network/virtualNetworks/virtualNetworkPeerings/write"
-		  ]
-		}
-	  }
-
-	  resource "azurerm_role_assignment" "assignment" {
-		principal_id       = azuread_service_principal.principal.id
-		scope              = azurerm_virtual_network.vnet.id
-		role_definition_id = azurerm_role_definition.definition.role_definition_resource_id
-	  }
-	  `, uniqueAzurePeeringTestID, subscriptionID, tenantID)
+	uniqueAzurePeeringTestID = fmt.Sprintf("hcp-provider-test-%s", time.Now().Format("200601021504"))
+	subscriptionID           = os.Getenv("ARM_SUBSCRIPTION_ID")
+	tenantID                 = os.Getenv("ARM_TENANT_ID")
 )
+
+var peeringHubSpokeNVAConfig = `
+	  allow_forwarded_traffic = true
+	  use_remote_gateways     = false
+`
+
+var peeringHubSpokeGatewayConfig = `
+	  allow_forwarded_traffic = true
+	  use_remote_gateways     = false
+`
+
+var peeringHubSpokeNVAandGatewayConfig = `
+	  allow_forwarded_traffic = true
+	  use_remote_gateways     = true
+`
+
+var fullConfig = baseConfig("", `
+	resource "azuread_service_principal" "principal" {
+	  application_id = hcp_azure_peering_connection.peering.application_id
+	}
+
+	resource "azurerm_role_definition" "definition" {
+	  name  = "hcp-provider-test-role-def"
+	  scope = azurerm_virtual_network.vnet.id
+
+	  assignable_scopes = [
+		azurerm_virtual_network.vnet.id
+	  ]
+
+	  permissions {
+		actions = [
+		  "Microsoft.Network/virtualNetworks/peer/action",
+		  "Microsoft.Network/virtualNetworks/virtualNetworkPeerings/read",
+		  "Microsoft.Network/virtualNetworks/virtualNetworkPeerings/write"
+		]
+	  }
+	}
+
+	resource "azurerm_role_assignment" "assignment" {
+	  principal_id       = azuread_service_principal.principal.id
+	  scope              = azurerm_virtual_network.vnet.id
+	  role_definition_id = azurerm_role_definition.definition.role_definition_resource_id
+	}
+`)
+
+func baseConfig(hubSpokeConfig, optConfig string) string {
+	return fmt.Sprintf(`
+	provider "azurerm" {
+	  features {}
+	}
+
+	resource "hcp_hvn" "test" {
+	  hvn_id         = "%[1]s"
+	  cloud_provider = "azure"
+	  region         = "eastus"
+	  cidr_block     = "172.25.16.0/20"
+	}
+
+	// This resource initially returns in a Pending state, because its application_id is required to complete acceptance of the connection.
+	resource "hcp_azure_peering_connection" "peering" {
+	  hvn_link                 = hcp_hvn.test.self_link
+	  peering_id               = "%[1]s"
+	  peer_vnet_name           = azurerm_virtual_network.vnet.name
+	  peer_subscription_id     = "%[2]s"
+	  peer_tenant_id           = "%[3]s"
+	  peer_resource_group_name = azurerm_resource_group.rg.name
+	  peer_vnet_region         = "eastus"
+
+	  // Hub/Spoke networking config
+      %[4]s
+	}
+
+	// This data source is the same as the resource above, but waits for the connection to be Active before returning.
+	data "hcp_azure_peering_connection" "peering" {
+	  hvn_link                 = hcp_hvn.test.self_link
+	  peering_id               = hcp_azure_peering_connection.peering.peering_id
+	  wait_for_active_state    = true
+	}
+
+	// The route depends on the data source, rather than the resource, to ensure the peering is in an Active state.
+	resource "hcp_hvn_route" "route" {
+	  hvn_route_id = "%[1]s"
+	  hvn_link = hcp_hvn.test.self_link
+	  destination_cidr = "172.31.0.0/16"
+	  target_link = data.hcp_azure_peering_connection.peering.self_link
+	}
+
+	resource "azurerm_resource_group" "rg" {
+	  name     = "%[1]s"
+	  location = "East US"
+	}
+
+	resource "azurerm_virtual_network" "vnet" {
+	  name                = "%[1]s"
+	  location            = azurerm_resource_group.rg.location
+	  resource_group_name = azurerm_resource_group.rg.name
+
+	  address_space = [
+		"10.0.0.0/16"
+	  ]
+	}
+
+	%[5]s
+	`, uniqueAzurePeeringTestID, subscriptionID, tenantID, hubSpokeConfig, optConfig)
+}
 
 func TestAccAzurePeeringConnection(t *testing.T) {
 	resourceName := "hcp_azure_peering_connection.peering"
@@ -119,7 +141,7 @@ func TestAccAzurePeeringConnection(t *testing.T) {
 		Steps: []resource.TestStep{
 			{
 				// Tests create
-				Config: testConfig(testAccAzurePeeringConfig),
+				Config: testConfig(fullConfig),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckAzurePeeringExists(resourceName),
 					resource.TestCheckResourceAttr(resourceName, "peering_id", uniqueAzurePeeringTestID),
@@ -157,7 +179,88 @@ func TestAccAzurePeeringConnection(t *testing.T) {
 			},
 			// Tests read
 			{
-				Config: testConfig(testAccAzurePeeringConfig),
+				Config: testConfig(fullConfig),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckAzurePeeringExists(resourceName),
+					resource.TestCheckResourceAttr(resourceName, "peering_id", uniqueAzurePeeringTestID),
+					testLink(resourceName, "hvn_link", uniqueAzurePeeringTestID, HvnResourceType, resourceName),
+					resource.TestCheckResourceAttr(resourceName, "peer_subscription_id", subscriptionID),
+					resource.TestCheckResourceAttr(resourceName, "peer_tenant_id", tenantID),
+					resource.TestCheckResourceAttr(resourceName, "peer_vnet_name", uniqueAzurePeeringTestID),
+					resource.TestCheckResourceAttrSet(resourceName, "peer_vnet_region"),
+					resource.TestCheckResourceAttrSet(resourceName, "azure_peering_id"),
+					resource.TestCheckResourceAttrSet(resourceName, "organization_id"),
+					resource.TestCheckResourceAttrSet(resourceName, "project_id"),
+					resource.TestCheckResourceAttrSet(resourceName, "created_at"),
+					resource.TestCheckResourceAttrSet(resourceName, "expires_at"),
+					resource.TestCheckResourceAttrSet(resourceName, "state"),
+					testLink(resourceName, "self_link", uniqueAzurePeeringTestID, PeeringResourceType, "hcp_hvn.test"),
+				),
+			},
+		},
+	})
+}
+
+// TestAccAzurePeeringConnectionInternal is almost identical to the standard
+// TestAccAzurePeeringConnection test, but does not include the Azure Service Principal
+// components which allow permission for HCP to peer from the "Customer" account
+// to HCP dataplane.
+// This modified test exists for HashiCorp internal contributors to adhere to on-dmand
+// service principal creation via doormat.
+func TestAccAzurePeeringConnectionInternal(t *testing.T) {
+	resourceName := "hcp_azure_peering_connection.peering"
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t, map[string]bool{"aws": false, "azure": true}) },
+		ProtoV6ProviderFactories: testProtoV6ProviderFactories,
+		ExternalProviders: map[string]resource.ExternalProvider{
+			"azurerm": {VersionConstraint: "~> 3.63"},
+			"azuread": {VersionConstraint: "~> 2.39"},
+		},
+		CheckDestroy: testAccCheckAzurePeeringDestroy,
+
+		Steps: []resource.TestStep{
+			{
+				// Tests create
+				Config: testConfig(baseConfig("", "")),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckAzurePeeringExists(resourceName),
+					resource.TestCheckResourceAttr(resourceName, "peering_id", uniqueAzurePeeringTestID),
+					testLink(resourceName, "hvn_link", uniqueAzurePeeringTestID, HvnResourceType, resourceName),
+					resource.TestCheckResourceAttr(resourceName, "peer_subscription_id", subscriptionID),
+					resource.TestCheckResourceAttr(resourceName, "peer_tenant_id", tenantID),
+					resource.TestCheckResourceAttr(resourceName, "peer_vnet_name", uniqueAzurePeeringTestID),
+					resource.TestCheckResourceAttrSet(resourceName, "allow_forwarded_traffic"),
+					resource.TestCheckResourceAttrSet(resourceName, "use_remote_gateways"),
+					resource.TestCheckResourceAttrSet(resourceName, "peer_vnet_region"),
+					resource.TestCheckResourceAttrSet(resourceName, "organization_id"),
+					resource.TestCheckResourceAttrSet(resourceName, "project_id"),
+					resource.TestCheckResourceAttrSet(resourceName, "created_at"),
+					resource.TestCheckResourceAttrSet(resourceName, "expires_at"),
+					resource.TestCheckResourceAttrSet(resourceName, "state"),
+					testLink(resourceName, "self_link", uniqueAzurePeeringTestID, PeeringResourceType, "hcp_hvn.test"),
+					// Note: azure_peering_id is not set until the peering is accepted after creation.
+				),
+			},
+			// Tests import
+			{
+				ResourceName: resourceName,
+				ImportState:  true,
+				ImportStateIdFunc: func(s *terraform.State) (string, error) {
+					rs, ok := s.RootModule().Resources[resourceName]
+					if !ok {
+						return "", fmt.Errorf("not found: %s", resourceName)
+					}
+
+					hvnID := s.RootModule().Resources["hcp_hvn.test"].Primary.Attributes["hvn_id"]
+					peerID := rs.Primary.Attributes["peering_id"]
+					return fmt.Sprintf("%s:%s", hvnID, peerID), nil
+				},
+				ImportStateVerify: true,
+			},
+			// Tests read
+			{
+				Config: testConfig(baseConfig("", "")),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckAzurePeeringExists(resourceName),
 					resource.TestCheckResourceAttr(resourceName, "peering_id", uniqueAzurePeeringTestID),
